requirements.txt
# Core data science packages
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Machine learning
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Web scraping and API
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.10.0
lxml>=4.9.0

# Database and data handling
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pymongo>=4.3.0

# Jupyter and development
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=6.5.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
click>=8.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Type checking
mypy>=1.5.0
types-requests>=2.31.0
